inventory-ring: fix label for single-item rings
