+++ src/tr1/game/inventory.c
@@ -2,14 +2,8 @@
 
 #include "game/gun.h"
 #include "game/inventory_ring/vars.h"
-#include "game/objects/common.h"
 #include "game/objects/vars.h"
-#include "global/types.h"
 #include "global/vars.h"
-
-#include <libtrx/utils.h>
-
-#include <stdint.h>
 
 bool Inv_AddItem(const GAME_OBJECT_ID obj_id)
 {
@@ -52,62 +46,32 @@ bool Inv_AddItem(const GAME_OBJECT_ID obj_id)
 
     case O_SHOTGUN_ITEM:
     case O_SHOTGUN_OPTION:
-        for (int32_t i = Inv_RequestItem(O_SHOTGUN_AMMO_ITEM); i > 0; i--) {
-            Inv_RemoveItem(O_SHOTGUN_AMMO_ITEM);
-            Inv_AddAmmo(&g_Lara.shotgun_ammo, SHOTGUN_AMMO_QTY);
-        }
-        Inv_AddAmmo(&g_Lara.shotgun_ammo, SHOTGUN_AMMO_QTY);
-        Inv_InsertItem(&g_InvRing_Item_Shotgun);
-        Item_GlobalReplace(O_SHOTGUN_ITEM, O_SHOTGUN_AMMO_ITEM);
+        Inv_AddGun(LGT_SHOTGUN);
         return false;
 
     case O_MAGNUM_ITEM:
     case O_MAGNUM_OPTION:
-        for (int32_t i = Inv_RequestItem(O_MAGNUM_AMMO_ITEM); i > 0; i--) {
-            Inv_RemoveItem(O_MAGNUM_AMMO_ITEM);
-            Inv_AddAmmo(&g_Lara.magnum_ammo, MAGNUM_AMMO_QTY);
-        }
-        Inv_AddAmmo(&g_Lara.magnum_ammo, MAGNUM_AMMO_QTY);
-        Inv_InsertItem(&g_InvRing_Item_Magnums);
-        Item_GlobalReplace(O_MAGNUM_ITEM, O_MAGNUM_AMMO_ITEM);
+        Inv_AddGun(LGT_MAGNUMS);
         return false;
 
     case O_UZI_ITEM:
     case O_UZI_OPTION:
-        for (int32_t i = Inv_RequestItem(O_UZI_AMMO_ITEM); i > 0; i--) {
-            Inv_RemoveItem(O_UZI_AMMO_ITEM);
-            Inv_AddAmmo(&g_Lara.uzi_ammo, UZI_AMMO_QTY);
-        }
-        Inv_AddAmmo(&g_Lara.uzi_ammo, UZI_AMMO_QTY);
-        Inv_InsertItem(&g_InvRing_Item_Uzis);
-        Item_GlobalReplace(O_UZI_ITEM, O_UZI_AMMO_ITEM);
+        Inv_AddGun(LGT_UZIS);
         return false;
 
     case O_SHOTGUN_AMMO_ITEM:
     case O_SHOTGUN_AMMO_OPTION:
-        if (Inv_RequestItem(O_SHOTGUN_ITEM)) {
-            Inv_AddAmmo(&g_Lara.shotgun_ammo, SHOTGUN_AMMO_QTY);
-        } else {
-            Inv_InsertItem(&g_InvRing_Item_ShotgunAmmo);
-        }
+        Inv_AddAmmo(LGT_SHOTGUN);
         return false;
 
     case O_MAGNUM_AMMO_ITEM:
     case O_MAGNUM_AMMO_OPTION:
-        if (Inv_RequestItem(O_MAGNUM_ITEM)) {
-            Inv_AddAmmo(&g_Lara.magnum_ammo, MAGNUM_AMMO_QTY);
-        } else {
-            Inv_InsertItem(&g_InvRing_Item_MagnumAmmo);
-        }
+        Inv_AddAmmo(LGT_MAGNUMS);
         return false;
 
     case O_UZI_AMMO_ITEM:
     case O_UZI_AMMO_OPTION:
-        if (Inv_RequestItem(O_UZI_ITEM)) {
-            Inv_AddAmmo(&g_Lara.uzi_ammo, UZI_AMMO_QTY);
-        } else {
-            Inv_InsertItem(&g_InvRing_Item_UziAmmo);
-        }
+        Inv_AddAmmo(LGT_UZIS);
         return false;
 
     case O_SMALL_MEDIPACK_ITEM:
