+++ src/tr2/game/inventory.c
@@ -1,6 +1,5 @@
 #include "game/inventory.h"
 
-#include "game/game_flow.h"
 #include "game/gun.h"
 #include "game/inventory_ring.h"
 #include "game/objects/vars.h"
@@ -8,83 +7,13 @@
 #include "global/vars.h"
 
 #include <libtrx/game/game.h>
-#include <libtrx/utils.h>
 
 static int32_t M_GetFlareQuantity(void);
-static INVENTORY_ITEM *M_GetGunInvItem(LARA_GUN_TYPE gun_type);
-static INVENTORY_ITEM *M_GetAmmoInvItem(LARA_GUN_TYPE gun_type);
-static void M_AddGun(LARA_GUN_TYPE gun_type);
 
 static int32_t M_GetFlareQuantity(void)
 {
     return Game_IsBonusFlagSet(GBF_JAPANESE) ? FLARE_AMMO_JAPANESE_QTY
                                              : FLARE_AMMO_QTY;
-}
-
-static INVENTORY_ITEM *M_GetGunInvItem(const LARA_GUN_TYPE gun_type)
-{
-    // clang-format off
-    switch (gun_type) {
-    case LGT_SHOTGUN: return &g_InvRing_Item_Shotgun;
-    case LGT_MAGNUMS: return &g_InvRing_Item_Magnums;
-    case LGT_UZIS:    return &g_InvRing_Item_Uzis;
-    case LGT_HARPOON: return &g_InvRing_Item_Harpoon;
-    case LGT_M16:     return &g_InvRing_Item_M16;
-    case LGT_GRENADE: return &g_InvRing_Item_Grenade;
-    default:          return nullptr;
-    }
-    // clang-format on
-}
-
-static INVENTORY_ITEM *M_GetAmmoInvItem(const LARA_GUN_TYPE gun_type)
-{
-    // clang-format off
-    switch (gun_type) {
-    case LGT_SHOTGUN: return &g_InvRing_Item_ShotgunAmmo;
-    case LGT_MAGNUMS: return &g_InvRing_Item_MagnumAmmo;
-    case LGT_UZIS:    return &g_InvRing_Item_UziAmmo;
-    case LGT_HARPOON: return &g_InvRing_Item_HarpoonAmmo;
-    case LGT_M16:     return &g_InvRing_Item_M16Ammo;
-    case LGT_GRENADE: return &g_InvRing_Item_GrenadeAmmo;
-    default:          return nullptr;
-    }
-    // clang-format on
-}
-
-static void M_AddGun(const LARA_GUN_TYPE gun_type)
-{
-    const GAME_OBJECT_ID gun_object = Gun_GetGunObject(gun_type);
-    const GAME_OBJECT_ID ammo_object = Gun_GetAmmoObject(gun_type);
-    AMMO_INFO *const ammo = Gun_GetAmmoInfo(gun_type);
-    const int32_t ammo_qty = Gun_GetAmmoQuantity(gun_type);
-    for (int32_t i = Inv_RequestItem(ammo_object); i > 0; i--) {
-        Inv_RemoveItem(ammo_object);
-        Inv_AddAmmo(ammo, ammo_qty);
-    }
-    Inv_AddAmmo(ammo, ammo_qty);
-    Inv_InsertItem(M_GetGunInvItem(gun_type));
-    if (g_Lara.last_gun_type == LGT_UNARMED) {
-        g_Lara.last_gun_type = gun_type;
-    }
-    if (Gun_IsRifleType(gun_type)) {
-        if (g_Lara.back_gun_obj_id == O_LARA) {
-            g_Lara.back_gun_obj_id = Gun_GetWeaponAnim(gun_type);
-            g_Lara.back_gun_type = gun_type;
-        }
-    }
-    Item_GlobalReplace(gun_object, ammo_object);
-}
-
-static void M_AddAmmo(const LARA_GUN_TYPE gun_type)
-{
-    const GAME_OBJECT_ID gun_object = Gun_GetGunObject(gun_type);
-    const int32_t ammo_qty = Gun_GetAmmoQuantity(gun_type);
-    AMMO_INFO *const ammo = Gun_GetAmmoInfo(gun_type);
-    if (Inv_RequestItem(gun_object)) {
-        Inv_AddAmmo(ammo, ammo_qty);
-    } else {
-        Inv_InsertItem(M_GetAmmoInvItem(gun_type));
-    }
 }
 
 bool Inv_AddItem(const GAME_OBJECT_ID obj_id)
@@ -138,62 +67,62 @@ bool Inv_AddItem(const GAME_OBJECT_ID obj_id)
 
     case O_SHOTGUN_ITEM:
     case O_SHOTGUN_OPTION:
-        M_AddGun(LGT_SHOTGUN);
+        Inv_AddGun(LGT_SHOTGUN);
         return false;
 
     case O_MAGNUM_ITEM:
     case O_MAGNUM_OPTION:
-        M_AddGun(LGT_MAGNUMS);
+        Inv_AddGun(LGT_MAGNUMS);
         return false;
 
     case O_UZI_ITEM:
     case O_UZI_OPTION:
-        M_AddGun(LGT_UZIS);
+        Inv_AddGun(LGT_UZIS);
         return false;
 
     case O_HARPOON_ITEM:
     case O_HARPOON_OPTION:
-        M_AddGun(LGT_HARPOON);
+        Inv_AddGun(LGT_HARPOON);
         return false;
 
     case O_M16_ITEM:
     case O_M16_OPTION:
-        M_AddGun(LGT_M16);
+        Inv_AddGun(LGT_M16);
         return false;
 
     case O_GRENADE_ITEM:
     case O_GRENADE_OPTION:
-        M_AddGun(LGT_GRENADE);
+        Inv_AddGun(LGT_GRENADE);
         return false;
 
     case O_SHOTGUN_AMMO_ITEM:
     case O_SHOTGUN_AMMO_OPTION:
-        M_AddAmmo(LGT_SHOTGUN);
+        Inv_AddAmmo(LGT_SHOTGUN);
         return false;
 
     case O_MAGNUM_AMMO_ITEM:
     case O_MAGNUM_AMMO_OPTION:
-        M_AddAmmo(LGT_MAGNUMS);
+        Inv_AddAmmo(LGT_MAGNUMS);
         return false;
 
     case O_UZI_AMMO_ITEM:
     case O_UZI_AMMO_OPTION:
-        M_AddAmmo(LGT_UZIS);
+        Inv_AddAmmo(LGT_UZIS);
         return false;
 
     case O_HARPOON_AMMO_ITEM:
     case O_HARPOON_AMMO_OPTION:
-        M_AddAmmo(LGT_HARPOON);
+        Inv_AddAmmo(LGT_HARPOON);
         return false;
 
     case O_M16_AMMO_ITEM:
     case O_M16_AMMO_OPTION:
-        M_AddAmmo(LGT_M16);
+        Inv_AddAmmo(LGT_M16);
         return false;
 
     case O_GRENADE_AMMO_ITEM:
     case O_GRENADE_AMMO_OPTION:
-        M_AddAmmo(LGT_GRENADE);
+        Inv_AddAmmo(LGT_GRENADE);
         return false;
 
     case O_SMALL_MEDIPACK_ITEM:
